docker-images/app/requirements.txt
adagio==0.2.4
aiosignal==1.3.1
annotated-types==0.5.0
antlr4-python3-runtime==4.11.1
anyio==3.7.1
appdirs==1.4.4
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.5
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.6
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.4
debugpy==1.6.7
decorator==5.1.1
docopt==0.6.2
exceptiongroup==1.1.2
executing==1.2.0
faiss-gpu==1.7.2
fastapi==0.101.0
filelock==3.9.0
frozenlist==1.4.0
fs==2.4.16
fsspec==2023.6.0
fugue==0.8.6
fugue-sql-antlr==0.1.6
grpcio==1.56.2
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.0
ipython==8.14.0
ipywidgets==8.1.0
jedi==0.19.0
Jinja2==3.1.2
json-numpy==1.0.1
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mpmath==1.2.1
msgpack==1.0.5
nest-asyncio==1.5.6
networkx==3.0
numpy==1.25.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
pip==23.2.1
pipreqs==0.4.13
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.0
pure-eval==0.2.2
pyarrow==12.0.1
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
pyzmq==23.2.1
qpd==0.4.4
referencing==0.30.2
requests==2.28.1
rpds-py==0.9.2
setuptools==68.0.0
six==1.16.0
sniffio==1.3.0
sqlglot==17.9.1
stack-data==0.6.2
starlette==0.27.0
sympy==1.11.1
tornado==6.2
traitlets==5.9.0
triad==0.9.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.13
uvicorn==0.23.2
wcwidth==0.2.6
wheel==0.38.4
widgetsnbextension==4.0.8
yarg==0.1.9
zipp==3.16.2
python-multipart
